refactor(tests): Split addTask tests
