+++ tests/unit/task-manager.test.js
@@ -169,46 +169,6 @@ const testParsePRD = async (prdPath, outputPath, numTasks) => {
 	}
 };
 
-// Create a simplified version of addTask for testing
-const testAddTask = (
-	tasksData,
-	taskPrompt,
-	dependencies = [],
-	priority = 'medium'
-) => {
-	// Create a new task with a higher ID
-	const highestId = Math.max(...tasksData.tasks.map((t) => t.id));
-	const newId = highestId + 1;
-
-	// Create mock task based on what would be generated by AI
-	const newTask = {
-		id: newId,
-		title: `Task from prompt: ${taskPrompt.substring(0, 20)}...`,
-		description: `Task generated from: ${taskPrompt}`,
-		status: 'pending',
-		dependencies: dependencies,
-		priority: priority,
-		details: `Implementation details for task generated from prompt: ${taskPrompt}`,
-		testStrategy: 'Write unit tests to verify functionality'
-	};
-
-	// Check dependencies
-	for (const depId of dependencies) {
-		const dependency = tasksData.tasks.find((t) => t.id === depId);
-		if (!dependency) {
-			throw new Error(`Dependency task ${depId} not found`);
-		}
-	}
-
-	// Add task to tasks array
-	tasksData.tasks.push(newTask);
-
-	return {
-		updatedData: tasksData,
-		newTask
-	};
-};
-
 // Import after mocks
 import * as taskManager from '../../scripts/modules/task-manager.js';
 import { sampleClaudeResponse } from '../fixtures/sample-claude-response.js';
@@ -316,56 +276,6 @@ describe('Task Manager Module', () => {
 			// 2. It reports errors for individual tasks
 			// 3. It completes the operation for successful tasks
 			expect(true).toBe(true);
-		});
-	});
-
-	describe('addTask function', () => {
-		test('should add a new task using AI', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-			const prompt = 'Create a new authentication system';
-
-			// Act
-			const result = testAddTask(testTasksData, prompt);
-
-			// Assert
-			expect(result.newTask.id).toBe(
-				Math.max(...sampleTasks.tasks.map((t) => t.id)) + 1
-			);
-			expect(result.newTask.status).toBe('pending');
-			expect(result.newTask.title).toContain(prompt.substring(0, 20));
-			expect(testTasksData.tasks.length).toBe(sampleTasks.tasks.length + 1);
-		});
-
-		test('should validate dependencies when adding a task', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-			const prompt = 'Create a new authentication system';
-			const validDependencies = [1, 2]; // These exist in sampleTasks
-
-			// Act
-			const result = testAddTask(testTasksData, prompt, validDependencies);
-
-			// Assert
-			expect(result.newTask.dependencies).toEqual(validDependencies);
-
-			// Test invalid dependency
-			expect(() => {
-				testAddTask(testTasksData, prompt, [999]); // Non-existent task ID
-			}).toThrow('Dependency task 999 not found');
-		});
-
-		test('should use specified priority', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-			const prompt = 'Create a new authentication system';
-			const priority = 'high';
-
-			// Act
-			const result = testAddTask(testTasksData, prompt, [], priority);
-
-			// Assert
-			expect(result.newTask.priority).toBe(priority);
 		});
 	});
